% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_processing.R
\name{assign_compound_classes}
\alias{assign_compound_classes}
\title{Assign compound classes using the chosen binset}
\usage{
assign_compound_classes(dat, BINSET)
}
\arguments{
\item{dat}{Input dataframe. This could be spectral data, or peak picked data.
Must include a `ppm` column for compound class assignment}

\item{A}{binset; e.g. \code{\link{bins_Clemente2012}},
\code{\link{bins_Hertkorn2013}}, etc., or a similarly-structured data frame}
}
\value{
A dataframe with columns describing the group name
(sometimes abbreviated), start and stop boundaries, and a longer, more
complete description of the group.
}
\description{
Use this function to import multiple spectra files, combine them,
and then process/clean the data.
}
