% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bins_Mitchell2018}
\alias{bins_Mitchell2018}
\title{NMR grouping bins from Mitchell et al. (2018).}
\format{
A data frame with 7 rows and 5 variables:
\describe{
  \item{number}{Bin number}
  \item{group}{Name of bin group}
  \item{start}{KP_TODO}
  \item{stop}{KP_TODO}
  \item{description}{Description of the bin group}
}
}
\source{
P Mitchell et al. 2018. “Nuclear Magnetic Resonance Analysis of
Changes in Dissolved Organic Matter Composition with Successive Layering
on Clay Mineral Surfaces.” Soil Systems
\url{https://doi.org/10.3390/soils2010008}.
}
\usage{
bins_Mitchell2018
}
\description{
NMR grouping bins from Mitchell et al. (2018). KP_TODO
}
\note{
The NMR spectrum can be split into several bins, based on
chemical shift (ppm). Binsets are specific to nuclei and solvents.
}
\keyword{datasets}
