% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_relabund.R
\name{compute_relabund_cores}
\alias{compute_relabund_cores}
\title{Assign compound classes using the chosen binset}
\usage{
compute_relabund_cores(SPECTRA_WITH_BINS, METHOD)
}
\arguments{
\item{SPECTRA_WITH_BINS}{Processed spectral data, output from `import_nmr_spectra_data` and `assign_compound_classes`}

\item{METHOD}{Choose the method for calculating relative abundance. Options include (a) "AUC", integrating the spectral region within each bin;
(b) "peaks", adding areas of peaks if a peak-picked file is provided.}
}
\value{
The output will be a dataframe with columns describing ...
}
\description{
Use this function to compute relative abundance of compound classes for each sample.
}
\examples{

}
\references{

}
