% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_relabund.R
\name{compute_relabund_treatments}
\alias{compute_relabund_treatments}
\title{Compute relative abundance summary by treatment}
\usage{
compute_relabund_treatments(RELABUND_CORES, TREATMENTS, COREKEY)
}
\arguments{
\item{RELABUND_CORES}{Dataframe with relative abundance for each sample (sampleID)}

\item{TREATMENTS}{columns being used for grouping to summarize the data}

\item{COREKEY}{Dataframe containing sample key}
}
\value{
The output will be a dataframe with columns describing ...
}
\description{
Use this function to compute relative-abundance summaries for each treatment
}
