% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_graphs.R
\name{gg_spectra}
\alias{gg_spectra}
\title{Compute relative abundance for each sample}
\usage{
gg_spectra(dat, BINSET, LABEL_POSITION, mapping, STAGGER)
}
\arguments{
\item{dat}{Processed spectral data, output from (a) \code{\link{nmr_import_spectra}}
and \code{\link{nmr_assign_bins}}; or (b) \code{\link{nmr_import_peaks}}}

\item{BINSET}{spectral binning}

\item{LABEL_POSITION}{y-axis position for bin labels}

\item{mapping}{aes(x = , y = )}

\item{STAGGER}{how much to stagger the labels? KP_TO (not clear)}
}
\value{
A dataframe with columns describing ... KP_TODO
}
\description{
Compute relative abundance of compound classes for each sample
and plot NMR spectra, with line-brackets denoting binned regions.
}
