% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_processing.R
\name{import_nmr_spectra_data}
\alias{import_nmr_spectra_data}
\title{Import and process NMR spectral data}
\usage{
import_nmr_spectra_data(path, method, pattern = "*.csv$", quiet = FALSE)
}
\arguments{
\item{path}{Directory where the spectra files are saved}

\item{method}{KP_TODO}

\item{pattern}{Filename pattern to search for (by default "*.csv$")}

\item{quiet}{Print diagnostic messages? Logical}
}
\value{
The data from all files found, concatenated into a single
data frame and sorted.
}
\description{
Imports multiple spectra files and then combines and cleans the data.
}
