% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_graphs.R
\name{nmr_plot_spectra}
\alias{nmr_plot_spectra}
\title{Plot NMR spectra}
\usage{
nmr_plot_spectra(dat, binset, label_position, mapping, stagger)
}
\arguments{
\item{dat}{Processed spectral data, output from (a) \code{\link{nmr_import_spectra}}
and \code{\link{nmr_assign_bins}}; or (b) \code{\link{nmr_import_peaks}}}

\item{binset}{spectral binning}

\item{label_position}{y-axis position for bin labels}

\item{mapping}{aes(x = , y = )}

\item{stagger}{how much to stagger the labels? KP_TO (not clear)}
}
\value{
A ggplot object
}
\description{
Plot NMR spectra, with line-brackets denoting binned regions.
Use spectra data processed in MestreNova or TopSpin.
}
\examples{
sdir <- system.file("extdata", "kfp_hysteresis", "spectra_mnova", package = "nmrrr")
spec <- nmr_import_spectra(path = sdir, method = "mnova")
library(ggplot2)
p_aes <- aes(x = ppm, y = intensity)
p <- nmr_plot_spectra(spec, bins_Clemente2012, 5, p_aes, STAGGER = 0.5)
p + ylim(0, 6)

tdir <- system.file("extdata", "meb_burn", "spectra_topspin", package = "nmrrr")
spec <- nmr_import_spectra(path = tdir, method = "topspin")
p_aes <- aes(x = ppm, y = intensity, color = sampleID)
p <- nmr_plot_spectra(spec, bins_Hertkorn2013, 6e+06, p_aes, STAGGER = 1e+06)
p + ylim(0, 8e+06)
}
