% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_processing.R
\name{process_peaks}
\alias{process_peaks}
\title{Process picked peaks data}
\usage{
process_peaks(PEAKS_FILES, METHOD)
}
\arguments{
\item{PEAKS_FILES}{file path for peaks data (input). All the peaks files are
saved as individual .csv files}

\item{METHOD}{KP_TODO}
}
\value{
The output will be a dataframe with columns describing
  sample ID, ppm, intensity, area, group name.
}
\description{
Use this function to process data of peaks picked with NMR software.
}
