% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bins_Hertkorn2013}
\alias{bins_Hertkorn2013}
\title{NMR grouping bins from Hertkorn et al. (2013).}
\format{
A data frame with 5 rows and 5 variables:
\describe{
  \item{number}{Bin number}
  \item{group}{Name of bin group}
  \item{start}{KP_TODO}
  \item{stop}{KP_TODO}
  \item{description}{Description of the bin group}
}
}
\source{
N. Hertkorn et al. 2013. "High-field NMR spectroscopy and FTICR mass
spectrometry: powerful discovery tools for the molecular level
characterization of marine dissolved organic matter" Biogeosciences
\url{https://doi.org/10.5194/bg-10-1583-2013}
}
\usage{
bins_Hertkorn2013
}
\description{
NMR grouping bins from  Hertkorn et al. (2013). KP_TODO
}
\note{
The NMR spectrum can be split into several bins, based on
chemical shift (ppm). Binsets are specific to nuclei and solvents.
}
\keyword{datasets}
