% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_graphs.R
\name{gg_spectra}
\alias{gg_spectra}
\title{Compute relative abundance for each sample}
\usage{
gg_spectra(dat, BINSET, LABEL_POSITION, mapping, STAGGER)
}
\arguments{
\item{dat}{Processed spectral data, output from (a) `import_nmr_spectra_data` and `assign_compound_classes`; or (b) process_peaks}

\item{BINSET}{spectral binning}

\item{LABEL_POSITION}{y-axis position for bin labels}

\item{mapping}{aes(x = , y = )}

\item{STAGGER}{how much to stagger the labels?}
}
\value{
The output will be a dataframe with columns describing ...
}
\description{
Compute relative abundance of compound classes for each sample.
This function will create a plot of NMR spectra, with line-brackets denoting binned regions
}
