% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_processing.R
\name{import_nmr_spectra_data}
\alias{import_nmr_spectra_data}
\title{Import and process NMR spectral data}
\usage{
import_nmr_spectra_data(SPECTRA_FILES, METHOD)
}
\arguments{
\item{SPECTRA_FILES}{path/directory where the spectra files are saved}
}
\value{
The output will be a dataframe with columns describing
  the group name (sometimes abbreviated), start and stop boundaries, and a longer, more complete description of the group.
}
\description{
Use this function to import multiple spectra files, combine them,
and then process/clean the data.
}
\examples{

}
\references{

}
