% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_processing.R
\name{import_nmr_spectra_data}
\alias{import_nmr_spectra_data}
\title{Import and process NMR spectral data}
\usage{
import_nmr_spectra_data(path, method, pattern = "*.csv$")
}
\arguments{
\item{path}{Directory where the spectra files are saved}

\item{method}{KP_TODO}

\item{pattern}{Filename pattern (by default "*.csv$")}
}
\value{
A dataframe with columns describing the group name (sometimes
abbreviated), start and stop boundaries, and a longer, more complete
description of the group.
}
\description{
Imports multiple spectra files and then combines and cleans the data.
}
