% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_graphs.R
\name{nmr_plot_spectra}
\alias{nmr_plot_spectra}
\title{Plot NMR spectra}
\usage{
nmr_plot_spectra(dat, BINSET, LABEL_POSITION, mapping, STAGGER)
}
\arguments{
\item{dat}{Processed spectral data, output from (a) \code{\link{nmr_import_spectra}}
and \code{\link{nmr_assign_bins}}; or (b) \code{\link{nmr_import_peaks}}}

\item{BINSET}{spectral binning}

\item{LABEL_POSITION}{y-axis position for bin labels}

\item{mapping}{aes(x = , y = )}

\item{STAGGER}{how much to stagger the labels? KP_TO (not clear)}
}
\value{
A ggplot object
}
\description{
Plot NMR spectra, with line-brackets denoting binned regions.
Use spectra data processed in MestreNova or TopSpin.
}
\examples{
sdir <- system.file("extdata", "kfp_hysteresis", "spectra_mnova", package = "nmrrr")
spec <- nmr_import_spectra(path = sdir, method = "mnova")
nmr_plot_spectra(spec, bins_Clemente2012, LABEL_POSITION = 5, aes(x = ppm, y = intensity), STAGGER = 0.5)+ ylim(0, 6)
tdir <- system.file("extdata", "meb_burn", "spectra_topspin", package = "nmrrr")
spec <- nmr_import_spectra(path = tdir, method = "topspin")
nmr_plot_spectra(spec, bins_Hertkorn2013, LABEL_POSITION = 6e+06, aes(x = ppm, y = intensity, color = sampleID), STAGGER = 1e+06)+ ylim(0, 8e+06)

}
