% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_relabund.R
\name{nmr_relabund}
\alias{nmr_relabund}
\title{Compute relative abundance for each sample}
\usage{
nmr_relabund(dat, method)
}
\arguments{
\item{dat}{Processed spectral data, output from (a) \code{\link{nmr_import_spectra}}
and \code{\link{nmr_assign_bins}}; or (b) \code{\link{nmr_import_peaks}}}

\item{method}{The method for calculating relative abundance.
Options include (a) "AUC", integrating the spectral region within each bin;
(b) "peaks", adding areas of peaks if a peak-picked file is provided.}
}
\value{
A dataframe with columns describing ... KP_TODO
}
\description{
Compute relative abundance of compound classes for each sample.
}
\examples{
tdir <- system.file("extdata", "meb_burn", "spectra_topspin", package = "nmrrr")
spec <- nmr_import_spectra(path = tdir, method = "topspin")
spec <- nmr_assign_bins(spec, bins_Hertkorn2013)
nmr_relabund(spec, "AUC")

pdir <- system.file("extdata", "amp_tempest", "peaks_mnova_single", package = "nmrrr")
peaks <- nmr_import_peaks(path = pdir, method = "single column")
peaks <- nmr_assign_bins(peaks, bins_Clemente2012)
nmr_relabund(peaks, "peaks")
}
