% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-functions_processing.R
\name{process_peaks}
\alias{process_peaks}
\title{Process picked peaks data}
\usage{
process_peaks(PEAKS_FILES, METHOD, BINSET)
}
\arguments{
\item{PEAKS_FILES}{file path for peaks data (input). All the peaks files are saved as individual .csv files}

\item{METHOD}{format of input data, depending on how the data were exported. "multiple columns": use when data are in split-column format, obtained by pasting "peaks table" in MNova.
"single column": use when data are in single-column format, exported from MNova as "peaks script".}

\item{BINSET}{Choose the binset you want. Options include: "Clemente2012", "Lynch2019", "Hertkorn2013_MeOD", "Mitchell2018"}
}
\value{
The output will be a dataframe with columns describing
  sample ID, ppm, intensity, area, group name.
}
\description{
Use this function to process data of peaks picked with NMR software.
}
\examples{

}
